+++ src/providers/self/recvmemorywindow.hpp
@@ -16,7 +16,7 @@
 #define pMR_PROVIDERS_SELF_RECVMEMORYWINDOW_H
 
 #include <memory>
-#include <cstdint>
+#include "config.hpp"
 
 namespace pMR { namespace self
 {
@@ -26,7 +26,7 @@ namespace pMR { namespace self
     {
         public:
             RecvMemoryWindow(std::shared_ptr<Connection> const,
-                    void *buffer, std::uint32_t const sizeByte);
+                    void *buffer, std::size_t const sizeByte);
             RecvMemoryWindow(const RecvMemoryWindow&) = delete;
             RecvMemoryWindow(RecvMemoryWindow&&) = delete;
             RecvMemoryWindow& operator=(const RecvMemoryWindow&) = delete;
@@ -38,7 +38,7 @@ namespace pMR { namespace self
         private:
             std::shared_ptr<Connection> const mConnection;
             void *mBuffer = nullptr;
-            std::uint32_t const mSizeByte;
+            std::size_t const mSizeByte;
     };
 }}
 #endif // pMR_PROVIDERS_SELF_RECVMEMORYWINDOW_H
