+++ src/providers/self/sendmemorywindow.cpp
@@ -17,12 +17,12 @@
 
 pMR::self::SendMemoryWindow::SendMemoryWindow(
         std::shared_ptr<Connection> const connection,
-        void *buffer, std::uint32_t const sizeByte)
+        void *buffer, std::size_t const sizeByte)
     :   mConnection(connection), mBuffer(buffer), mSizeByte{sizeByte} { }
 
 void pMR::self::SendMemoryWindow::init() { }
 
-void pMR::self::SendMemoryWindow::post(std::uint32_t const sizeByte)
+void pMR::self::SendMemoryWindow::post(std::size_t const sizeByte)
 {
     mConnection->pollAddress();
     mConnection->sendData(mBuffer, {sizeByte});
